module5-solution
Summary: In this assignment, we are going to have a bit of fun with our
built restaurant web application. The front page of our web app
contains 3 clickable tiles: Menu, Specials, and Map. If you click on
the Specials tile, you will be taken to a single category page where
all the menu items that belong to the Specials menu category will be
shown. Your task in this assignment is to alter this behavior such that
when the user clicks on the Specials tile, the web app takes the user
to a random single category menu page, listing menu items in the
category, be it "Lunch", "Dinner", "Sushi", etc. That way, any random
category can become the Specials! What fun! (not! :-) )